Relocate dataCall and connect keywords to header via props

Relates #4
